nginx-flask-mysql/backend/requirements.txt
Flask==2.0.1
mysql-connector==2.2.9
Requests==2.31.0